[backend] Moved app creation to createMainServer
